+++ src/Commands/0_core/ping.js
@@ -2,6 +2,7 @@ const Command = require('../../Structures/Command');
 
 const { emoji: { success } } = require('../../../config/config.json');
 
+
 module.exports = new Command({
 	name: 'ping',
 	aliases: [ 'pong' ],