+++ src/Commands/0_core/say.js
@@ -6,7 +6,7 @@ const { emoji: { warning }, response: { missingArguments } } = require('../../..
 module.exports = new Command({
 	name: 'say',
 	aliases: [ ' ' ],
-	syntax: 'say [message]',
+	syntax: 'say <message>',
 	description: 'Repeats whatever shit you said, then quietly deletes your message',
 	async run(message, args, client) {
 		if (!args[0]) return message.channel.send(`${warning} ${missingArguments}`);