+++ src/Commands/1_sound/stop.js
@@ -2,7 +2,7 @@ const Command = require('../../Structures/Command.js');
 
 const { Permissions } = require('discord.js');
 const { getVoiceConnection } = require('@discordjs/voice');
-const { emoji: { success, warning }, response: { wrongChannel, noMusic } } = require('../../../config/config.json');
+const { bot: { ownerID }, emoji: { success, warning }, response: { wrongChannel, noMusic } } = require('../../../config/config.json');
 
 
 module.exports = new Command({
@@ -13,7 +13,7 @@ module.exports = new Command({
 		const connection = getVoiceConnection(message.guild.id);
 		if (!connection) return message.channel.send(`${warning} ${noMusic}`);
 
-		if (( !message.member.voice.channel || connection.joinConfig.channelId != message.member.voice.channel.id ) && !message.member.permissions.has(Permissions.FLAGS.ADMINISTRATOR)) return message.channel.send(`${warning} ${wrongChannel}`);
+		if (( !message.member.voice.channel || connection.joinConfig.channelId != message.member.voice.channel.id ) && !message.member.permissions.has(Permissions.FLAGS.MANAGE_CHANNELS) && message.author.id != ownerID) return message.channel.send(`${warning} ${wrongChannel}`);
 
 		connection.destroy();
 		message.channel.send(`'${success} The player has stopped.`);