+++ src/Data/settings.js
@@ -11,6 +11,15 @@ const checkSettingsFiles = () => {
         }
         writeFileSync('./config/settings.json', JSON.stringify(data, null, 4));
     }
+    else { // Check if the settings file supports the new settings
+        const settingsFile = require('../../config/settings.json');
+        Object.entries(settingsFile.guild).forEach(([key, value]) => {
+            if (value.enabledDefaultJoin === undefined) {
+                value.enabledDefaultJoin = true;
+            }
+        });
+        writeFileSync('./config/settings.json', JSON.stringify(settingsFile, null, 4));
+    }
     fileLog('[INFO] Settings file checked!');
 }